moved notifications code to programs
